% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxDot.R
\name{MaxDot}
\alias{MaxDot}
\title{Treatment Position Vs Treatment Scatter Plot}
\usage{
MaxDot(design)
}
\arguments{
\item{design}{Input a design in matrix format}
}
\value{
Generates scatter plot of treatment position Vs treatment
}
\description{
The function will give the scatter plot showing the spread offered by design points in the experimental space. The x axis of the plot represent treatments and y axis the position of treatments in each block. Different colors in the dot represent the blocks.
}
\examples{
library(SFOCDs)
mat<-matrix(c(  1,  4,  2,  5,
2,  5,  3,  6,
3,  6,  1,  4,
4,  1,  5,  2,
5,  2,  6,  3,
6,  3,  4,  1),nrow=6,byrow=TRUE)
MaxDot(mat)
}
