% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maxpro_Criterion.R
\name{Maxpro_Criterion}
\alias{Maxpro_Criterion}
\title{Maxpro Criterion}
\usage{
Maxpro_Criterion(design)
}
\arguments{
\item{design}{Input a design in matrix format}
}
\value{
Maxpro criterion value for a given design
}
\description{
Maxpro Criterion
}
\examples{
library(SFOCDs)
mat<-matrix(c(
1,2,3,
2,1,4,
3,4,1,
4,3,2),nrow=4,byrow=TRUE)
Maxpro_Criterion(mat)
}
