% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshuffle_des.R
\name{reshuffle_des}
\alias{reshuffle_des}
\title{Reshuffle Design}
\usage{
reshuffle_des(design, type = 1)
}
\arguments{
\item{design}{Input any base design}

\item{type}{Input 1 or 2. By default set 1. If type =2 then reshuffled for JuxtaOCDs.}
}
\value{
Reshuffled the base design such that each column has minimum possible repeatations of symbols.
}
\description{
Reshuffle Design
}
\examples{
library(SFOCDs)
mat<-matrix(c(1,2,3,4,1,2,4,5,1,3,4,5,2,3,4,5),nrow=4,byrow=TRUE)
reshuffle_des(mat,type=1)
}
